// src/providers/mystifly/flight-book-responses/dtos/request/update-mystifly-flight-book-response-request.dto.ts
export class UpdateMystiflyFlightBookResponseRequestDto {
  id: number;
  ClientUTCOffset: number;
  ConversationId: string;
  Errors: any;
  Status: string;
  TktTimeLimit: string;
  TraceId: string;
  UniqueID: string;
  providerReference: string;
  paymentReferenceNumber: string;
  sequence: number;
}